ReadStatus클래스 Channel의 도메인 밸류로 이동. Channel -> ChannelMember -> ReadStatus
