+++ src/main/java/com/sprint/mission/discodeit/application/BasicServerService.java
@@ -2,7 +2,7 @@ package com.sprint.mission.discodeit.application;
 
 
 import com.sprint.mission.discodeit.domain.readstatus.ReadStatusRepository;
-import com.sprint.mission.discodeit.domain.readstatus.ReadStatus;
+import com.sprint.mission.discodeit.domain.channel.ReadStatus;
 import com.sprint.mission.discodeit.domain.server.ServerRepository;
 import com.sprint.mission.discodeit.domain.server.Server;
 import com.sprint.mission.discodeit.application.dto.request.ServerCreateRequestDto;
