+++ src/main/java/com/sprint/mission/discodeit/domain/channel/Message.java
@@ -1,19 +1,21 @@
 package com.sprint.mission.discodeit.domain.channel;
 
+import com.sprint.mission.discodeit.domain.binarycontent.BinaryContent;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 
+import java.io.Serializable;
+import java.time.Instant;
 import java.util.UUID;
 
 @Getter
 @RequiredArgsConstructor
-public class Message {
-
-    private static final long serialVersionUID = 9L;
+public class Message implements Serializable {
 
     private final UUID senderId;
     private final String content;
-    private final UUID binaryContentId;
+    private final BinaryContent binaryContentId;
+    private final Instant createdAt;
 
 
 }
