--- src/main/java/com/sprint/mission/discodeit/domain/readstatus/ReadStatusRepository.java
@@ -1,16 +1,0 @@
-package com.sprint.mission.discodeit.domain.readstatus;
-
-import java.util.List;
-import java.util.Optional;
-import java.util.UUID;
-
-public interface ReadStatusRepository {
-
-    void save(ReadStatus readStatus);
-
-    void remove(ReadStatus readStatus);
-
-    Optional<ReadStatus> findById(UUID id);
-
-    List<ReadStatus> findAll();
-}
