--- src/main/java/com/sprint/mission/discodeit/infrastructure/jcf/JCFReadStatusRepository.java
@@ -1,37 +1,0 @@
-package com.sprint.mission.discodeit.infrastructure.jcf;
-
-import com.sprint.mission.discodeit.domain.readstatus.ReadStatusRepository;
-import com.sprint.mission.discodeit.domain.readstatus.ReadStatus;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
-import org.springframework.stereotype.Repository;
-
-import java.util.*;
-
-@Repository
-@ConditionalOnProperty(prefix = "discodeit.repository", name = "type", havingValue = "jcf")
-public class JCFReadStatusRepository implements ReadStatusRepository {
-
-    private final Map<UUID, ReadStatus> store = new HashMap<>();
-
-    @Override
-    public void save(ReadStatus readStatus) {
-        UUID key = readStatus.getId();
-        store.put(key,readStatus);
-    }
-
-    @Override
-    public void remove(ReadStatus readStatus) {
-        UUID key = readStatus.getId();
-        store.remove(key);
-    }
-
-    @Override
-    public Optional<ReadStatus> findById(UUID id) {
-        return Optional.ofNullable(store.get(id));
-    }
-
-    @Override
-    public List<ReadStatus> findAll() {
-        return List.copyOf(store.values());
-    }
-}
